requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.17.0
numpy==1.24.3
pillow==10.0.1
gunicorn==21.2.0
